Conflicts:
	l10ntools/Module_l10ntools.mk
	l10ntools/source/merge.cxx
	solenv/gbuild/AllLangResTarget.mk
	solenv/gbuild/ExtensionTarget.mk
	solenv/gbuild/InstallModuleTarget.mk
	solenv/gbuild/TargetLocations.mk

+++ package/qa/storages/TestHelper.java
@@ -98,7 +98,7 @@ public class TestHelper  {
         // check size property of the stream
         try
         {
-            int nSize = AnyConverter.toInt( xPropSet.getPropertyValue( "Size" ) );
+            long nSize = AnyConverter.toLong( xPropSet.getPropertyValue( "Size" ) );
             if ( nSize != pBytes.length )
             {
                 Error( "The 'Size' property of substream '" + sStreamName + "' contains wrong value!" );
@@ -188,7 +188,7 @@ public class TestHelper  {
         // check size property of the stream
         try
         {
-            int nSize = AnyConverter.toInt( xPropSet.getPropertyValue( "Size" ) );
+            long nSize = AnyConverter.toLong( xPropSet.getPropertyValue( "Size" ) );
             if ( nSize != pBytes.length )
             {
                 Error( "The 'Size' property of substream '" + sStreamName + "' contains wrong value!" );
@@ -825,7 +825,7 @@ public class TestHelper  {
             {
                 // get "MediaType" and "Size" properties and control there values
                 String sPropMediaType = AnyConverter.toString( xPropSet.getPropertyValue( "MediaType" ) );
-                int nPropSize = AnyConverter.toInt( xPropSet.getPropertyValue( "Size" ) );
+                long nPropSize = AnyConverter.toLong( xPropSet.getPropertyValue( "Size" ) );
                 boolean bPropCompress = AnyConverter.toBoolean( xPropSet.getPropertyValue( "Compressed" ) );
 
                 bOk = true;
