+++ vcl/aqua/source/gdi/salnativewidgets.cxx
@@ -709,7 +709,7 @@ sal_Bool AquaSalGraphics::drawNativeControl(ControlType nType,
                 // Perhaps some other HIThemeButtonDrawInfo setting would do the trick ?
 
                 CGContextSetShouldAntialias( mrContext, true );
-                float aGrey[] = { 0.45, 0.45, 0.45, 1.0 };
+                CGFloat aGrey[] = { 0.45, 0.45, 0.45, 1.0 };
                 CGContextSetFillColor( mrContext, aGrey );
                 CGContextBeginPath( mrContext );
                 float x = rc.origin.x + rc.size.width;
